Revert "esm (#65)"
This reverts commit c6bf12277c140285eb37270a9a064029fa7ada5c.
