+++ config/webpack.config.js
@@ -1,12 +1,10 @@
-import path from 'path';
+const path = require("path");
 
-const __dirname = path.dirname(new URL(import.meta.url).pathname);
-
-export default {
-  mode: 'development',
-  entry: './lib/index.js',
+module.exports = {
+  mode: "development",
+  entry: "./lib/index.js",
   output: {
-    path: path.resolve(__dirname, '../build'),
-    filename: 'index.js',
+    path: path.resolve(__dirname, "../build"),
+    filename: "index.js",
   },
 };
